adding information from frontend to backend
