+++ lifetracker-api/models/user.js
@@ -12,18 +12,14 @@ class User {
       username,
       first_name,
       last_name,
-      password,
-      created_at,
-      updated_at,
+      password
     } = creds;
     const requiredCreds = [
       "email",
       "password",
       "username",
       "first_name",
-      "last_name",
-      "created_at",
-      "updated_at",
+      "last_name"
     ];
     // try {
     //   validateFields({ required: requiredCreds, obj: creds, location: "user registration" })
@@ -34,11 +30,24 @@ class User {
     const existingUserWithEmail = await User.fetchUserByEmail(email);
     if (existingUserWithEmail) {
       throw new BadRequestError(`Duplicate email: ${email}`);
-      return false;
+    //   return false;
     }
+    console.log(existingUserWithEmail)
 
     const hashedPassword = await bcrypt.hash(password, BCRYPT_WORK_FACTOR);
     const normalizedEmail = email.toLowerCase();
+
+    const currentDate = new Date();
+    const year = currentDate.getFullYear();
+    const month = (currentDate.getMonth() + 1).toString().padStart(2, "0");
+    const date = currentDate.getDate().toString().padStart(2, "0");
+    const hours = currentDate.getHours().toString().padStart(2, "0");
+    const minutes = currentDate.getMinutes().toString().padStart(2, "0");
+    const seconds = currentDate.getSeconds().toString().padStart(2, "0");
+    const datetime = `${year}-${month}-${date} ${hours}:${minutes}:${seconds}`;
+
+    const created_at = datetime;
+    const updated_at = datetime;
 
     const result = await db.query(
       `INSERT INTO users (
@@ -64,7 +73,7 @@ class User {
     );
 
     const user = result.rows[0];
-    console.log(user)
+    console.log(user);
 
     return true;
   }
